fix(inquiry): correctly logs in relay environment
