+++ src/v2/Components/Inquiry/Hooks/useArtworkInquiryRequest.ts
@@ -1,55 +1,56 @@
-import { commitMutation, graphql, Environment } from "relay-runtime"
-import { useSystemContext } from "v2/System"
+import { commitMutation, graphql } from "relay-runtime"
 import {
   SubmitInquiryRequestMutationInput,
   useArtworkInquiryRequestMutation,
 } from "v2/__generated__/useArtworkInquiryRequestMutation.graphql"
+import { useInquiryContext } from "./useInquiryContext"
 
 type UseArtworkInquiryRequestInput = Omit<
   SubmitInquiryRequestMutationInput,
   "inquireableID" | "inquireableType" | "message"
 > & {
-  relayEnvironment?: Environment
   artworkID: string
   message: string
   contactGallery?: boolean | null
 }
 
 export const useArtworkInquiryRequest = () => {
-  const { relayEnvironment: defaultRelayEnvironment } = useSystemContext()
+  const { relayEnvironment } = useInquiryContext()
 
   const submitArtworkInquiryRequest = ({
-    relayEnvironment = defaultRelayEnvironment,
     artworkID,
     ...rest
   }: UseArtworkInquiryRequestInput) => {
     return new Promise((resolve, reject) => {
-      commitMutation<useArtworkInquiryRequestMutation>(relayEnvironment!, {
-        onCompleted: (res, errors) => {
-          if (errors !== null) {
-            reject(errors)
-            return
-          }
-
-          resolve(res)
-        },
-        mutation: graphql`
-          mutation useArtworkInquiryRequestMutation(
-            $input: SubmitInquiryRequestMutationInput!
-          ) {
-            submitInquiryRequestMutation(input: $input) {
-              clientMutationId
+      commitMutation<useArtworkInquiryRequestMutation>(
+        relayEnvironment.current!,
+        {
+          onCompleted: (res, errors) => {
+            if (errors !== null) {
+              reject(errors)
+              return
             }
-          }
-        `,
-        variables: {
-          input: {
-            inquireableID: artworkID,
-            inquireableType: "Artwork",
-            ...rest,
+
+            resolve(res)
           },
-        },
-      })
+          mutation: graphql`
+            mutation useArtworkInquiryRequestMutation(
+              $input: SubmitInquiryRequestMutationInput!
+            ) {
+              submitInquiryRequestMutation(input: $input) {
+                clientMutationId
+              }
+            }
+          `,
+          variables: {
+            input: {
+              inquireableID: artworkID,
+              inquireableType: "Artwork",
+              ...rest,
+            },
+          },
+        }
+      )
     })
   }
